split tcp/http modules isolated async IO fix qty_of_inverters in HTTP modules
